requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
tortoise-orm>=0.20.0
orjson>=3.9.10
polars>=0.20.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
asyncpg>=0.29.0 